requirements.txt
pillow
picamera
RPi.GPIO
psutil
tweepy